Setup project scaffold
This commit sets up the project scaffold depicted in the below diagram.
There are some typescript errors, but these will be smoothed in
subsequent PRs

Note that there is a lot left to do but running `npm run build-core`
will now create two separate builds: a browser (wasm) and a node native
module.