Introduce cross platform testing
This introduces the ability to run tests inside a nodejs or browser
environment. So when running in a browser environment, all the browser
APIs should be available (like Indexeddb or fetch, so no mocking
needed).

This was accomplished using a combination of jest and karmatic (karma +
webpack (rollup support is on master but not released)). Now one can
actually run a test and verify that these different environment receive
different implementations.

I switched up some of the top level API to make typescript happy --
though I'm not entirely happy with it, but I'm sure a better API will be
teased out as implementations are filled in.