Hook up N-API to in memory store
This commit hooks up the node native module so that the store functions
are actually callable and work as intended. In the interest of time,
this was accomplished by swapping napi-rs with node_bindgen. Both are
high quality projects but since node_binden has less ceremony I was able
to implement it quicker than if I had continued down the napi-rs path.

The good news is that all tests pass on both web and native
implementations :tada:

One important note important note is that for the native module, futures
are required to be `Send` (thread-safe), but the common store trait
disables this, so I figured it's not worth the hassle and dropped the
explicit implementation.